add hooks for, and proof of, using multi-entity modseq for complex state
We want a way to treat type X entities as changed if their associated
type Y entity is changed.  This commit adds some not-quite-fully-baked
abstractions for doing that, built on top of the last few commits, which
were in furtherance of the same goal.

The main complications are ensuring that the updates are fetched in a
consistent order so that getting updates with maxChanges can, in fact,
make progress toward a synchronized state without missing states.