egregious hack: _return_ix_get
This lets an rclass grab the to-be-returned results of an ix_get so that
it can tack on more stuff.  Ned uses this for fetchMembersForMe, but
it's awful.  The weird-o division of Result and ResultSet makes fine
sense in DBIC, but it's caused a fair bit of confusion in Ix.

I wonder whether what we want is to have Ix::DBIC::Entity, which
generates and installs both a result and a resultset, with all the
methods in the right place.  Hm!