save states after each ix_set
This allows multiple setFoos in one request to bump the state twice,
but it's not quite right.  If an internal client calls ix_update directly,
it won't work.  ix_* should be able to save state based on whether it's in
a larger grouping, possibly with some form of action deferral.  For now,
though this is good enough.