check for defined data_type without respect to db_data_type
The previous version of this check would allow the program to proceed
when db_data_type was set but data_type was not.  This would let the
DBIx::Class know the type, but ->ix_property_info would still have
no data_type in it.  This check requires that data_type itself is set.