is_nullable must be given a defined false
Due to some bits of crazy in DBIx-Class and SQL-Translator, providing
(is_nullable=>undef) will be treated like the default value for
SQL-Translator, which is true -- even though the default value for DBIC
is false.

I've done some lobbying for change on DBIC's part, but for now this
fixes things for us.