refactor the JMAP handling bits of Ix::App to its own method
The goal here is to let Ix::App-derived classes handle requests that
aren't the JMAP handled by the Ix::Processor without giving up the
ix.transaction logging system.

With this refactoring, someone can override _core_request and, if they
want, provide their own context and response.  The transaction will
still be logged.

Note that this replaces the ix.transaction's "calls" with "body,"
since calls are specific to JMAP.  Loggers will have to be updated for
that change.

Obviously this is all even more experimental than most bits of Ix.