do not promote null sort or filter to containers
Right now, sort and filter might be autovivified from undef to []
and {} respectively.  This preserves the original value.

If acting on the filters alters their contents rather than just
promoting them, this is not enough, but we should be mapping new
values into a new structure, not altering the input in place.

Tests for this are found in a downstream application.