ix_set - lock the state row for the collection being updated
This ensures that updates are atomic for a collection.

We no longer have an overarching transaction around a set of JMAP method
calls.

Instead, we start one transaction inside of ix_set around set/update/destroy
so that a single method call holds a lock and is released when the method
completes.

We also refresh our view of the current account states between method calls
incase they have changed out from under us (or we thought they changed but
dismissed the changes when a method call failed somehow).