getFooListUpdates - query on all immutable fields instead of required fields
Some classes may not have any required filters at all (at least, not in
metadata).

To determine which subset of entites we need to look at, we can query
those properties that they passed in that are immutable.

For now, immutable means:

 * not in the list of ix_mutable_properties
 * OR specifies an xref_to

I believe all xref_tos are really immutable, but we don't have a way
to actually make them so in metadata yet. (is_immutable will prevent
rows from being created with those properties passed in).

We really need is_initializable for this.