Catch exceptions in main app after database changes have committed
After we've committed database changes, it's possible to throw an exception
in log_access/log_transaction and friends. If that happens, we do *not* want
to return a 500 as we already have a prepared good response for the user.

Instead, we should attempt to report these new exceptions, and then return
the original response.