Fail to compile if classes have filters that cannot be diffed
If a class specifies a filter that doesn't:

 - match an ix property on the class
 - provide its own differ subroutine
 - refer to a joined relationship

then fail to compile mentioning why.

This saves consumers trouble down the road if they implement
a filter that doesn't match up to an Ix property but don't provide
a custom differ. When this would happen, Ix would crash during
ix_get_list_updates() when trying to compare rows against filters.