Foo/changes: return error if delta > maxChanges
The old behavior was: if the smallest delta we could compute was more
than the client's maxChanges request, we would give you a single state's
worth of changes. A spec change (from October 2016!) means that we
should actually return a cannotCalculateChanges error in this case
instead, because the client must always be able to prevent itself from
being flooded.